LayoutTests/fast/sub-pixel/float-with-right-margin-zoom-expected.txt

PASS elements[0].getBoundingClientRect().top is elements[1].getBoundingClientRect().top
Both elements below should be on the same line.
